requirements.txt
ruff==0.4.4
boto3==1.34.113
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1